Resolve OKR-1 "Feature/ fix objective api ver 2"

See merge request aibles-spring-learning/graduation-thesis/okrs/okrs-backend!72

+++ okrs-core/src/main/java/org/ptit/okrs/core/entity/Objective.java
@@ -28,9 +28,9 @@ public class Objective extends BaseEntityWithUpdater {
   private String description;
 
   @Column(nullable = false)
-  private Long startDate;
+  private Integer startDate;
 
-  private Long endDate;
+  private Integer endDate;
 
   @Enumerated(EnumType.STRING)
   private OkrsType type;
